refactor(agent): split fallback extraction out of XmlOutputManager
XmlOutputManager.ts had grown to ~880 lines with three unrelated concerns
inlined at module level. Move the fallback-extraction machinery into
focused modules under src/agent/output/extraction/:

- responseText.ts    — thinking-tag stripping + line normalization (deduped
                       between the header and fence collectors)
- markdownFences.ts  — fence delimiter parsing and stripping
- latexHeuristics.ts — LaTeX document-shape heuristics
- filenameHeaders.ts — %-header / bare-label document recovery (now a pure
                       function; the class passes labelFiles and
                       singleInputName explicitly)
- contentSimilarity.ts — unlabeled-fence collection and greedy
                       similarity assignment

No behavior changes; XmlOutputManager keeps the tier pipeline, XML parsing,
and file writing (414 lines).
